+++ vyper/ast/__init__.py
@@ -1,0 +1,1 @@
+from .utils import parse_to_ast  # noqa: F401
